==> code/cleaning_scripts/02a_merge_2011_candidate_with_treat.R <==
rm(list=ls())

library(tidyverse)
library(sf)
library(readxl)
library(lubridate)
library(numform)
library(here)


# READ IN ASSASSINATIONS DATA ------------------
viol <- read_excel(here("data", "raw_data", "assassinations_data", "za_assassinations copy.xlsx"))


# CLEAN ASSASSINATIONS DATA ------------------
viol$year <- year(viol$date_of_attack)
viol$month <- format(viol$date_of_attack, "%m")
viol$month <- f_num(viol$month, digits=0)
viol$sitting <- as.numeric(viol$sitting)
viol$electoral_cycle <- as.integer(viol$electoral_cycle)

# FILTER TO LOOK EXCLUSIVELY AT ATTACKS ON LC/MC WARD POLITICIANS AND/OR CANDIDATES ------------------
viol_ward <- viol %>% 
  filter(seat_type=="LC Ward" | seat_type=="MC Ward" | seat_type=="MC Ward candidate" | seat_type=="LC Ward candidate") %>% 
  filter(date_of_attack>="2000-12-05" & date_of_attack<="2016-08-03") 

viol_ward_sitting <- viol_ward %>% 
  filter(sitting == 1)

viol_2006_to_2011 <- viol_ward_sitting %>% 
  filter(date_of_attack > "2006-03-01" & date_of_attack < "2011-05-18")

# MANUALLY LINK 2006 WARD IDS TO 2016 WARD IDS USING CENTROID COORDINATES ------------------


# first we will need to read in 2006 and 2016 geometries
  # read in 2006 geomtry
ward2006 <- st_read(here("data", "gis_data_raw", "Wards2006.shp"))
ward2006 <- ward2006 %>% 
  rename(ward_id = WARD_ID)

ward2006 <- st_transform(ward2006, 6148) #transform to SA's official coordinate system, Hartebeesthoek94 (code 6148)



  # read in 2016 geometry
ward_2016 <- st_read(here("data", "gis_data_raw", "Wards2016.shp"))

ward_2016 <- st_transform(ward_2016, 6148) #transform to SA's official coordinate system, Hartebeesthoek94 (code 6148)
names(ward_2016)

ward_2016 <- ward_2016 %>% 
  select(-c(OBJECTID, Shape__Are, Shape__Len, DistrictMu, District_1, ProvinceNa, WardNumber, Year)) %>% 
  rename(province_id = ProvinceCo,
         local_municipality_id = LocalMunic,
         ward_id = WardID, 
         local_municipality_name = LocalMun_1)

  # transform to sf object 
ward_2016 <- st_as_sf(ward_2016)




#### WARD 1
print(viol_2006_to_2011$ward_id)
pnts1 <- ward2006 %>% 
  filter(ward_id == "52306006")

centroids1 <- as.data.frame(st_coordinates(st_centroid(pnts1$geometry)))

interpolate_centroids1 <- st_as_sf(centroids1, coords = c("X", "Y"), crs = st_crs(6148))

interpolate1 <- interpolate_centroids1 %>% mutate(
  intersection = as.integer(st_intersects(geometry, ward_2016)),
  area = if_else(is.na(intersection), '', ward_2016$ward_id[intersection])
)

pnts1$ward_id
interpolate1$area
  #so 52306006 is now 52307015


##### WARD 2
print(viol_2006_to_2011$ward_id)
pnts2 <- ward2006 %>% 
  filter(ward_id == "52901005")

centroids2 <- as.data.frame(st_coordinates(st_centroid(pnts2$geometry)))

interpolate_centroids2 <- st_as_sf(centroids2, coords = c("X", "Y"), crs = st_crs(6148))

interpolate2 <- interpolate_centroids2 %>% mutate(
  intersection = as.integer(st_intersects(geometry, ward_2016)),
  area = if_else(is.na(intersection), '', ward_2016$ward_id[intersection])
)

pnts2$ward_id
interpolate2$area
#so 52901005 is still 52901005

##### WARD 3
print(viol_2006_to_2011$ward_id)
pnts3 <- ward2006 %>% 
  filter(ward_id == "52405005")

centroids3 <- as.data.frame(st_coordinates(st_centroid(pnts3$geometry)))

interpolate_centroids3 <- st_as_sf(centroids3, coords = c("X", "Y"), crs = st_crs(6148))

interpolate3 <- interpolate_centroids3 %>% mutate(
  intersection = as.integer(st_intersects(geometry, ward_2016)),
  area = if_else(is.na(intersection), '', ward_2016$ward_id[intersection])
)

pnts3$ward_id
interpolate3$area
#so 52405005 is still 52405005

##### WARD 4
print(viol_2006_to_2011$ward_id)
pnts4 <- ward2006 %>% 
  filter(ward_id == "52405009")

centroids4 <- as.data.frame(st_coordinates(st_centroid(pnts4$geometry)))

interpolate_centroids4 <- st_as_sf(centroids4, coords = c("X", "Y"), crs = st_crs(6148))

interpolate4 <- interpolate_centroids4 %>% mutate(
  intersection = as.integer(st_intersects(geometry, ward_2016)),
  area = if_else(is.na(intersection), '', ward_2016$ward_id[intersection])
)

pnts4$ward_id
interpolate4$area
#so 52405009 is now 52405003


##### WARD 5
print(viol_2006_to_2011$ward_id)
pnts5 <- ward2006 %>% 
  filter(ward_id == "52804015")

centroids5 <- as.data.frame(st_coordinates(st_centroid(pnts5$geometry)))

interpolate_centroids5 <- st_as_sf(centroids5, coords = c("X", "Y"), crs = st_crs(6148))

interpolate5 <- interpolate_centroids5 %>% mutate(
  intersection = as.integer(st_intersects(geometry, ward_2016)),
  area = if_else(is.na(intersection), '', ward_2016$ward_id[intersection])
)

pnts5$ward_id
interpolate5$area
#so 52804015 is still 52804015

##### WARD 6
print(viol_2006_to_2011$ward_id)
pnts6 <- ward2006 %>% 
  filter(ward_id == "52603005")

centroids6 <- as.data.frame(st_coordinates(st_centroid(pnts6$geometry)))

interpolate_centroids6 <- st_as_sf(centroids6, coords = c("X", "Y"), crs = st_crs(6148))

interpolate6 <- interpolate_centroids6 %>% mutate(
  intersection = as.integer(st_intersects(geometry, ward_2016)),
  area = if_else(is.na(intersection), '', ward_2016$ward_id[intersection])
)

pnts6$ward_id
interpolate6$area
#so 52603005 is still 52603005

##### WARD 7
print(viol_2006_to_2011$ward_id)
pnts7 <- ward2006 %>% 
  filter(ward_id == "52401003")

centroids7 <- as.data.frame(st_coordinates(st_centroid(pnts7$geometry)))

interpolate_centroids7 <- st_as_sf(centroids7, coords = c("X", "Y"), crs = st_crs(6148))

interpolate7 <- interpolate_centroids7 %>% mutate(
  intersection = as.integer(st_intersects(geometry, ward_2016)),
  area = if_else(is.na(intersection), '', ward_2016$ward_id[intersection])
)

pnts7$ward_id
interpolate7$area
#so 52401003 is now 52401001


##### WARD 8
print(viol_2006_to_2011$ward_id)
pnts8 <- ward2006 %>% 
  filter(ward_id == "83007003")

centroids8 <- as.data.frame(st_coordinates(st_centroid(pnts8$geometry)))

interpolate_centroids8 <- st_as_sf(centroids8, coords = c("X", "Y"), crs = st_crs(6148))

interpolate8 <- interpolate_centroids8 %>% mutate(
  intersection = as.integer(st_intersects(geometry, ward_2016)),
  area = if_else(is.na(intersection), '', ward_2016$ward_id[intersection])
)

pnts8$ward_id
interpolate8$area
#so 83007003 is now 83007002


##### WARD 9
print(viol_2006_to_2011$ward_id)
pnts9 <- ward2006 %>% 
  filter(ward_id == "19100086")

centroids9 <- as.data.frame(st_coordinates(st_centroid(pnts9$geometry)))

interpolate_centroids9 <- st_as_sf(centroids9, coords = c("X", "Y"), crs = st_crs(6148))

interpolate9 <- interpolate_centroids9 %>% mutate(
  intersection = as.integer(st_intersects(geometry, ward_2016)),
  area = if_else(is.na(intersection), '', ward_2016$ward_id[intersection])
)

pnts9$ward_id
interpolate9$area
#so 19100086 is still 19100086


##### WARD 10
print(viol_2006_to_2011$ward_id)
pnts10 <- ward2006 %>% 
  filter(ward_id == "59200068")

centroids10 <- as.data.frame(st_coordinates(st_centroid(pnts10$geometry)))

interpolate_centroids10 <- st_as_sf(centroids10, coords = c("X", "Y"), crs = st_crs(6148))

interpolate10 <- interpolate_centroids10 %>% mutate(
  intersection = as.integer(st_intersects(geometry, ward_2016)),
  area = if_else(is.na(intersection), '', ward_2016$ward_id[intersection])
)

pnts10$ward_id
interpolate10$area
#so 59200068 is now 59500075


##### WARD 11
print(viol_2006_to_2011$ward_id)
pnts11 <- ward2006 %>% 
  filter(ward_id == "52402007")

centroids11 <- as.data.frame(st_coordinates(st_centroid(pnts11$geometry)))

interpolate_centroids11 <- st_as_sf(centroids11, coords = c("X", "Y"), crs = st_crs(6148))

interpolate11 <- interpolate_centroids11 %>% mutate(
  intersection = as.integer(st_intersects(geometry, ward_2016)),
  area = if_else(is.na(intersection), '', ward_2016$ward_id[intersection])
)

pnts11$ward_id
interpolate11$area
#so 52402007 is still 52402007


##### WARD 12
print(viol_2006_to_2011$ward_id)
pnts12 <- ward2006 %>% 
  filter(ward_id == "52902004")

centroids12 <- as.data.frame(st_coordinates(st_centroid(pnts12$geometry)))

interpolate_centroids12 <- st_as_sf(centroids12, coords = c("X", "Y"), crs = st_crs(6148))

interpolate12 <- interpolate_centroids12 %>% mutate(
  intersection = as.integer(st_intersects(geometry, ward_2016)),
  area = if_else(is.na(intersection), '', ward_2016$ward_id[intersection])
)

pnts12$ward_id
interpolate12$area
#so 52902004 is now 52902021


##### WARD 13
print(viol_2006_to_2011$ward_id)
pnts13 <- ward2006 %>% 
  filter(ward_id == "52304007")

centroids13 <- as.data.frame(st_coordinates(st_centroid(pnts13$geometry)))

interpolate_centroids13 <- st_as_sf(centroids13, coords = c("X", "Y"), crs = st_crs(6148))

interpolate13 <- interpolate_centroids13 %>% mutate(
  intersection = as.integer(st_intersects(geometry, ward_2016)),
  area = if_else(is.na(intersection), '', ward_2016$ward_id[intersection])
)

pnts13$ward_id
interpolate13$area
#so 52304007 is now 52307022


##### WARD 14
print(viol_2006_to_2011$ward_id)
pnts14 <- ward2006 %>% 
  filter(ward_id == "64005007")

centroids14 <- as.data.frame(st_coordinates(st_centroid(pnts14$geometry)))

interpolate_centroids14 <- st_as_sf(centroids14, coords = c("X", "Y"), crs = st_crs(6148))

interpolate14 <- interpolate_centroids14 %>% mutate(
  intersection = as.integer(st_intersects(geometry, ward_2016)),
  area = if_else(is.na(intersection), '', ward_2016$ward_id[intersection])
)

pnts14$ward_id
interpolate14$area
#so 64005007 is now 74804007


##### WARD 15
print(viol_2006_to_2011$ward_id)
pnts15 <- ward2006 %>% 
  filter(ward_id == "52105003")

centroids15 <- as.data.frame(st_coordinates(st_centroid(pnts15$geometry)))

interpolate_centroids15 <- st_as_sf(centroids15, coords = c("X", "Y"), crs = st_crs(6148))

interpolate15 <- interpolate_centroids15 %>% mutate(
  intersection = as.integer(st_intersects(geometry, ward_2016)),
  area = if_else(is.na(intersection), '', ward_2016$ward_id[intersection])
)

pnts15$ward_id
interpolate15$area
#so 52105003 is now 52106030


##### WARD 16
print(viol_2006_to_2011$ward_id)
pnts16 <- ward2006 %>% 
  filter(ward_id == "52405011")

centroids16 <- as.data.frame(st_coordinates(st_centroid(pnts16$geometry)))

interpolate_centroids16 <- st_as_sf(centroids16, coords = c("X", "Y"), crs = st_crs(6148))

interpolate16 <- interpolate_centroids16 %>% mutate(
  intersection = as.integer(st_intersects(geometry, ward_2016)),
  area = if_else(is.na(intersection), '', ward_2016$ward_id[intersection])
)

pnts16$ward_id
interpolate16$area
#so 52405011 is now 52405006



#NOW READ IN THE CLEANED CANDIDATES DATA AND CREATE A VARIABLE FOR ASSASSINATION TREATMENT
data_2011 <- readRDS(here("data", "processed_data", "candidates_clean_2011_extrapolated_to_2016.RDS"))


# CREATE DUMMY VARIABLE FOR ANY EXPOSED UNITS ------------------
data_2011 <- data_2011 %>% 
  mutate(treat = case_when(
    ward_id == "52307015" ~ 1,
    ward_id == "52901005" ~ 1,
    ward_id == "52405005" ~ 1,
    ward_id == "52405003" ~ 1,
    ward_id == "52804015" ~ 1,
    ward_id == "52603005" ~ 1,
    ward_id == "52401001" ~ 1,
    ward_id == "83007002" ~ 1,
    ward_id == "19100086" ~ 1,
    ward_id == "59500075" ~ 1,
    ward_id == "52402007" ~ 1,
    ward_id == "52902021" ~ 1,
    ward_id == "52307022" ~ 1,
    ward_id == "74804007" ~ 1,
    ward_id == "52106030" ~ 1,
    ward_id == "52405006" ~ 1
    )) %>% 
  mutate(target_party = case_when(
    ward_id == "52307015" ~ "IFP",
    ward_id == "52901005" ~ "IFP",
    ward_id == "52405005" ~ "IFP",
    ward_id == "52405003" ~ "IFP",
    ward_id == "52804015" ~ "IFP",
    ward_id == "52603005" ~ "ANC",
    ward_id == "52401001" ~ "IFP",
    ward_id == "83007002" ~ "ANC",
    ward_id == "19100086" ~ "ANC",
    ward_id == "59500075" ~ "Independent",
    ward_id == "52402007" ~ "IFP",
    ward_id == "52902021" ~ "ANC",
    ward_id == "52307022" ~ "IFP",
    ward_id == "74804007" ~ "ANC",
    ward_id == "52106030" ~ "ANC",
    ward_id == "52405006" ~ "IFP"
  )) %>% 
  mutate(target_name = case_when(
    ward_id == "52307015" ~ "Mcobiseni Patrick Duma",
    ward_id == "52901005" ~ "Simon Dingindawo Shange",
    ward_id == "52405005" ~ "Enoch Sibongiseni Shange",
    ward_id == "52405003" ~ "Emmanuel Ntuthuko Ngcobo",
    ward_id == "52804015" ~ "Mfanafuthi Elliot Maphumulo",
    ward_id == "52603005" ~ "Mfana Alpheus Xulu",
    ward_id == "52401001" ~ "Petrous Nxele",
    ward_id == "83007002" ~ "Thandi Pauline Mtsweni",
    ward_id == "19100086" ~ "Xolani Ronald Sotashe",
    ward_id == "59500075" ~ "Rajahrathanum Marimuthu Naidoo",
    ward_id == "52402007" ~ "Dumelani Jerome Zulu",
    ward_id == "52902021" ~ "Preva Khumalo",
    ward_id == "52307022" ~ "Dlangamandla Rich Mchunu",
    ward_id == "74804007" ~ "Mphela Dorah Modise",
    ward_id == "52106030" ~ "Maurus Zakhele Cele",
    ward_id == "52405006" ~ "Mnziwami Benard Mbatha"
  )) %>% 
  mutate(target_gender = case_when(
    ward_id == "52307015" ~ "male",
    ward_id == "52901005" ~ "male",
    ward_id == "52405005" ~ "male",
    ward_id == "52405003" ~ "male",
    ward_id == "52804015" ~ "male",
    ward_id == "52603005" ~ "male",
    ward_id == "52401001" ~ "male",
    ward_id == "83007002" ~ "female",
    ward_id == "19100086" ~ "male",
    ward_id == "59500075" ~ "male",
    ward_id == "52402007" ~ "male",
    ward_id == "52902021" ~ "male",
    ward_id == "52307022" ~ "male",
    ward_id == "74804007" ~ "female",
    ward_id == "52106030" ~ "male",
    ward_id == "52405006" ~ "male"
  )) %>% 
  mutate(local_municipality_id = case_when(
    ward_id == "52307015" ~ "KZN237",
    ward_id == "52901005" ~ "KZN291",
    ward_id == "52405005" ~ "KZN245",
    ward_id == "52405003" ~ "KZN245",
    ward_id == "52804015" ~ "KZN284",
    ward_id == "52603005" ~ "KZN263",
    ward_id == "52401001" ~ "KZN241",
    ward_id == "83007002" ~ "MP307",
    ward_id == "19100086" ~ "CPT",
    ward_id == "59500075" ~ "ETH",
    ward_id == "52402007" ~ "KZN242",
    ward_id == "52902021" ~ "529020",
    ward_id == "52307022" ~ "KZN292",
    ward_id == "74804007" ~ "GT484",
    ward_id == "52106030" ~ "KZN216",
    ward_id == "52405006" ~ "KZN245"
  ))


# check to make sure this worked properly: 
length(which(data_2011$treat==1)) ###16 woohoo!


# WRITE OUT DATA TO PROCESSED_DATA FILE
saveRDS(data_2011, here("data", "processed_data", "data_2011_with_assassinations_merged.rds"))
